Implement "nav lock selected wingman"
This required a little restructuring.  We can't let the code for which
wingman to follow be computed at compile time, since that's before we
know the value of `selected_wingman`.  It's a little ugly, but we can
put a special case in the code run when we push `APENG` and compute the
correct button to push *then*.